% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bmeta_stan.R
\name{bmeta_ana}
\alias{bmeta_ana}
\title{Risk difference or Risk Ratio}
\usage{
bmeta_ana(
  ctl_event,
  ctl_total,
  trt_event,
  trt_total,
  type = c("riskdiff", "riskratio"),
  tau_max = 2,
  tau2_eta = 100,
  ...
)
}
\description{
Risk difference or Risk Ratio
}
