\name{print.bmeta}
\alias{print.bmeta}
\title{
Print method for \code{bmeta} objects
}
\description{
Function to print output from function \code{bmeta}
}
\usage{
\method{print}{bmeta}(x, ...)
}
\arguments{
  \item{x}{
a \code{bmeta} object with results of the model
}
  \item{\dots}{
other arguments
}
}
\author{
Tao Ding
Gianluca Baio
}

