% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acf.plot.R
\name{acf.plot}
\alias{acf.plot}
\title{Autocorrelation function plot}
\usage{
acf.plot(x, node, title = "Autocorrelation function")
}
\arguments{
\item{x}{a \code{bmeta} object with results of the model}

\item{node}{variable to be displayed on the plot}

\item{title}{title of the plot, if specified}
}
\value{
A plot showing the autocorrelation for the selected node
}
\description{
Function to create autocorrelation function plot to assess convergence
}
\examples{

### Read and format the data (binary)
data = read.csv(url("https://gianluca.statistica.it/software/bmeta/Data-bin.csv"))

### List data for binary outcome 
data.list <- list(y0=data$y0,y1=data$y1,n0=data$n0,n1=data$n1) 

### generate output from bmeta 
x <- bmeta(data=data.list,outcome="bin",model="std.dt",type="ran")

### generate autocorrelation function plot 
acf.plot(x,"alpha[1]")

### generate autocorrelation function plot and specify the title 
acf.plot(x,"alpha[1]",title="Autocorrelation plot")

}
\author{
Tao Ding Gianluca Baio
}
\keyword{Autocorrelation}
\keyword{MCMC}
