% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.bmeta.R
\name{print.bmeta}
\alias{print.bmeta}
\title{Print method for \code{bmeta} objects}
\usage{
\method{print}{bmeta}(x, ...)
}
\arguments{
\item{x}{a \code{bmeta} object with results of the model}

\item{\dots}{other arguments}
}
\description{
Function to print output from function \code{bmeta}
}
\author{
Tao Ding Gianluca Baio
}
